Automatically restore older basebackup on repeated failure
It is possible for basebackups to be corrupt in some rare cases.
Depending on the issue it could possibly be resolved manually but
automatic restoration would fail. The binary logs in such cases should
be usable and restoring older basebackup and all binary logs from the
older backup and the new backup whose basebackup was unusable should
yield the same end result.