Allow gap in GTID sequence when patching GTID executed
Code assumed that the first binlog we're restoring has all the
transactions since the last persisted update to the gtid_executed table
because the table should be persisted whenever binlog is rotated. It
seems there are some cases where the backed up gtid_executed table has
not been updated as per that assumption. In the cases seen so far all
data was still present and patching GTID executed as usual would've been
the correct action.