Reschedule binary log prefetch operations for new restore coordinator
If the process is restarted or restore coordinator is re-created due to
config change it could end up in a state where it has pending binary
logs that haven't been downloaded but the download operations were not
scheduled so it ended up doing nothing.

Also log all scheduled prefetch operations and successful responses.