Fix restoring backup with no binary logs to restore
There should practically always be some binary logs to restore but when
doing point-in-time restore and setting restore target time before the
very first backup was created all binary logs will be excluded and that
caused MySQL server not to be started when some MyHoard logic expected
it to be. Add explicit check for whether it has been started before
trying to access it.