Do not fail promotion if slave status is missing
Slave status may be missing because someone has performed RESET SLAVE
(problem) or because there were zero binlogs to restore when the
restoring the service from backup (not a problem). We don't really know
which one is causing the state to be missing but typically people aren't
running RESET SLAVE so assume the latter.