daemon: Fix handling of simultaneous reload requests
If MyHoard got SIGHUP while it was still processing the earlier one it
could end up trying to create new WebServer when previous one was still
starting up, causing the WebServer initialization to fail. This resulted
in old WebServer remaining active with reference to old controller. That
WebServer instance returned obsolete data for backups and was unable to
serve requests like service promotion, causing the daemon to appear to
be functioning normally yet possibly failing at a critical time (when
promotion needed to be done).