Handle gtid_executed truncation correctly for multiple servers
In some cases it is possible that the active binary log contains
transactions from multiple different servers when a new backup is being
created. In such cases the logic that tried to ensure that the
transactions that are reported to be available in the basebackup do not
contain any transactions that have happened after the basebackup was
created didn't work correctly as it assumed the binary log only had
transactions from single server and didn't parse the string correctly.