Keep track of latest completed binlog when BackupStream was created
Binary logs that were created before a new backup was started but were
not yet purged could not be purged while basebackup for the new backup
was being created because the new stream didn't have remote references
for those binary logs. Binary logs that exist when new backup is started
cannot be required by the backup stream because basebackup must contain
everything in those so they can be simply ignored by the safety check.