Refresh backup streams less frequently on active node
Only active MyHoard node is expected to create new backup streams or
make changes to existing ones and whenever it does so it already updates
it's state immediately instead of relying on polling. The polling is
mostly just a safety mechanism to ensure changes are picked up at some
point if for some reason there are unexpected changes. This can be done
much less frequently than for non-active nodes.