Support servers configured with slave_parallel_workers > 0
Previously the restoration would fail if slave_parallel_workers was not
zero. Now parallel workers should work as well.

The main issue was that if not picking carefully enough the batch of
relay logs to apply when re-initializing the slaves the call to
`CHANGE MASTER TO` could end up permanently failing. This issue seemed
to happen if the relay logs contained no transactions at all or they
contained only transactions that had been already applied (and in some
cases if they just contained transactions that had been applied even if
there also were unapplied transactions). The batch selection logic has
been updated to ensure the batch contains new transactions (or then
treat the batch as the final one).

In case the SQL threads still go into bad state the entire slave state
is reset and process is continued from where it left. Because doing the
reset is unsafe if there are temporary tables there is explicit check
for that condition and the restoration just breaks if slave threads go
into bad state while temporary tables exist. This case did not happen
during testing but there are likely ways to make it happen. In such
event the only solution would be to configure the server not to use
parallel workers and restart the entire process.