Handle invalid basebackup binlog position
The binlog position reported by XtraBackup does not always match the
actual start position of the transaction following the last GTID
included in the backup but it may instead be the one after that. This
looks to be locking issue on MySQL side that causes inconsistent state
to be returned.

To work around the issue, add manual matching of the reported last GTID
to actual binlog position while restoring backup. It would perhaps be
better to fix this in the actual backup creation phase to avoid invalid
backup metadata but that would require much more complex logic and doing
this in restoration allows coping with old backups that are already
broken.