Automatically start SQL thread if not running
There have been some cases where SQL thread has been expected to be
running but it isn't. It's not clear why it has not started. It may
have been related to the the re-entrancy issues with
_apply_downloaded_remote_binlogs but logs don't clearly support that.
Log the expected replication target to make it easier to see what the
application is doing and if the thread isn't running try to start it
automatically.