Allow forcing premature server promotion
In case binary logs cannot be applied cleanly (they are corrupt or
contain changes that are too slow to apply), it is now possible to
use the POST /status API to forcibly mark restoration as having
completed and switch MyHoard to active mode. This option will always
cause data loss and must only be used when there are no other options
available.