myhoard: Add context manager for running BackupStream objects
This principally exists to make testing easier, where we need to control
the lifecycle of the objects via their threads and shutdown on arbitrary
exceptions.

The advantage is that when running tests, the cause of an error is made
clear since we don't cause a bunch of secondary errors when the state dirs
disappear from under the threads.