myhoard: Fix occasional failover problems
At some point the pymysql error returned for ER_MASTER_INFO changed to be
an OperationalError rather then an InternalError, which means that we
weren't handling this problem properly depending on version.

Add an extra exception catch to look for both types of problem, and react
appropriately. This would only have affected newer deployments.