disable flaky test and add ticket BF-1198
As soon as container images were added, this test was found to be flaky.
It seems to depend on some condition which wasn't triggered in Github
Actions native container but gets triggered very often within the
container.

This deserves an additional investigation, but such failures would be
red herrings right now.