test_util: proper test
By spawning bash, we had a race condition, because (of course) bash
would then fork another process which could match the
detect_running_process_id condition (since the inner command was
identical).

So, we launch just one non-forking child process,
and we rely on the uniqueness of three random integers.