Add tool to assist in monitoring progress
Because backups and restores consist of multiple phases,
provide a small helper utility to map progress in each phase onto
a single continuum.

In addition, support the concept of phases having multiple steps. This
is specifically designed to assist with binlog restores, where we
do not know in advance how many passes we will need before we
have caught up.