Reupload stream closed marker in case of state mismatch
Sometimes we end up in a situation where the closed.json for a stream
does not exist in object storage but local state says the stream has been
closed. This condition should not be possible because object storage is
updated before updating local state and any error in updating object
storage should cause the local state not to be updated.

To recover from this supposedly impossible situation, have controller
pass the closed timestamp read from object storage when it creates a
backup stream object and if backup stream sees that the object storage
closed marker is not present even though it should be then it uploads
it again, hopefully resolving the issue.