Add ability to delete individual backups
Currently backups are deleted only if they age out, or if there are too
many. There are situations where backups are taken of corrupted data
that need to be removed. The new /backups/{stream_id} endpoint supports
the GET and DELETE verbs as a natural extension of the existing /backups
collection endpoint. The backup still has to be closed before it can be
deleted.