Prevent myhoard starting more than one backup stream
Multiple backup streams add complexity and instability,
increasing load on the server for negligible benefit. It is
better to complete the current backup before devoting resources
to beginning a second stream.

This also has the side-effect of better handling situations where
one stream appears to fail, then recovers, and interferes with a
new backup stream: until the first one is finished, a second one
will not start, after which the first one is gone.