support backup preservation
Sometimes, it becomes necessary to retain a backup, as there may be instances
where a backup is in the process of being restored, but it suddenly gets removed due to its age
and the current amount of backups is surpassing the maximum allowed to be retained.

To prevent such scenario, one can request the preservation of a specific backup until a particular date.
To initiate this process, a PUT request to /backup/{stream_id}/preserve must be executed.
This action will prompt the controller to add the preservation request under 'pending_preservation_requests,'
signifying that it will subsequently update the backup's preservation status while in 'active' mode.

The preservation of a backup involves the storage of a 'preserve.json' file within the site's object storage, which contains the designated 'preserve until' date."

[BF-2219]