Fix error taking a backup using PXB 8.0.35
Latest versions of PXB support only one type of stream and
it should be specified using `=`. This fix makes it work for both 8.0.30
and recent versions.

Also decreased expected value of bytes read in the test for the backup,
it seems the default compression algorithm produces smaller file.

Fixed counting of restored files: qpress files are not supported
anymore. Default files use zstd format and end with .ztd.xbcrypt.