Ensure MySQL is only restarted once before starting streaming changes
In some rare cases MySQL takes a long time to start accepting queries.
When that happens, the server may take several minutes to do some
initial processing, after which it starts responding normally again and
subsequent restarts are fast. Restarting it before it has finished the
processing causes the operation to never complete so avoid doing that
so that the system does not go to never ending restart loop.